use 32 bit pointers

+++ libc/include/stddef.h
@@ -1,10 +1,12 @@
 #pragma once
 
+#include <stdint.h>
+
 #define NULL ((void*)0)
 
-typedef unsigned short size_t;
-typedef signed short ptrdiff_t;
+typedef uint32_t size_t;
+typedef int32_t ptrdiff_t;
 
 #define offsetof(type, member) ((size_t)&(((type*)0)->member))
 
-typedef signed short ssize_t; // actually not defined in standard, but useful
+typedef int32_t ssize_t; // actually not defined in standard, but useful
