+++ libc/include/stdint.h
@@ -2,14 +2,14 @@
 
 typedef signed char int8_t;
 typedef signed short int16_t;
-typedef signed long int32_t;
+typedef signed int int32_t;
 
 typedef unsigned char uint8_t;
 typedef unsigned short uint16_t;
-typedef unsigned long uint32_t;
+typedef unsigned int uint32_t;
 
-typedef int16_t intptr_t;
-typedef uint16_t uintptr_t;
+typedef int32_t intptr_t;
+typedef uint32_t uintptr_t;
 
 #define INT16_MAX 32767
 #define INT16_MIN -32768
